feat: add latch mode and hold -> latch mode switcher
```
So there is already a Hold to record. hotkey I want to add another latching hotkey and this is if I press AltGr twice It
should start recording and when I press AltGr, it should stop recording

Another thing that I want to introduce is If I'm holding AltGr to record, i should be able to press space to go into this
latch mode. And when I press AltGr, it should stop recording.

These actions will apply to right Cmd key as well.

Plan and think how you would go about implementing this by looking at the code. And then give me a plan So I could
approve it.
```